Extract common code to fetch cluster mappings
and use it in Datastores and Networks
